CDH-77566. Better scaling of k8s apps with dynamic allocation.
This change implements two changes to the k8s scheduler to behave
a little better when dynamic allocation is on.

The first allows the application to ramp up immediately when there's
a change in the target number of executors. Without this change,
scaling would only trigger when a change happened in the state of
the cluster, e.g. an executor going down.

The second gets rid of pending pod requests, both acknowledged (i.e.
Spark knows that a pod is pending resource allocation) and unacknowledged
(i.e. Spark has requested the pod but the API server hasn't created it
yet), when they're not needed anymore. This avoids starting those
executors to just remove them after the idle timeout, wasting resources
in the meantime.

Change-Id: I120db47cbd7f2fb074f596130f2eac8d84095c60