CDH-77734.  Check java version w/out commons-lang3
We need to perform some checks based on the jdk version, but
unfortunately updating the version of commons-lang3 introduces some
classpath issues b/c hive-exec bundles its own commons-lang3 which is
really old.  This causes problems in (a) our unit tests and (b) oozie.

So we just directly check the java version ourself, if we have a problem
with the commons-lang3 check.

(cherry picked from commit d6394b7)

Change-Id: I57574d54b0214c312ed8a102acafac1f0dbb3b66